# artifacts/api-server/src/routes/orders.ts
import { Router } from "express";
import { db } from "@workspace/db";
import { ordersTable, orderItemsTable, productsTable, usersTable, assetsTable, inventoryItemsTable, inventoryMovementsTable } from "@workspace/db";
import { eq, and, inArray } from "drizzle-orm";
import { requireAuth, requireTenant } from "../lib/auth";
import { writeAuditLog } from "../lib/audit";

const router = Router();

async function buildOrderResponse(o: typeof ordersTable.$inferSelect) {
  const items = await db.select({
    id: orderItemsTable.id,
    productId: orderItemsTable.productId,
    productName: productsTable.name,
    productNameAr: productsTable.nameAr,
    quantity: orderItemsTable.quantity,
    unitPrice: orderItemsTable.unitPrice,
    totalPrice: orderItemsTable.totalPrice,
    notes: orderItemsTable.notes,
  }).from(orderItemsTable)
    .leftJoin(productsTable, eq(orderItemsTable.productId, productsTable.id))
    .where(eq(orderItemsTable.orderId, o.id));

  let assetName = null, assetNameAr = null, createdByUserName = null, assignedToUserName = null;
  if (o.assetId) {
    const [a] = await db.select({ name: assetsTable.name, nameAr: assetsTable.nameAr }).from(assetsTable).where(eq(assetsTable.id, o.assetId)).limit(1);
    assetName = a?.name ?? null; assetNameAr = a?.nameAr ?? null;
  }
  if (o.createdByUserId) {
    const [u] = await db.select({ name: usersTable.name }).from(usersTable).where(eq(usersTable.id, o.createdByUserId)).limit(1);
    createdByUserName = u?.name ?? null;
  }
  if (o.assignedToUserId) {
    const [u] = await db.select({ name: usersTable.name }).from(usersTable).where(eq(usersTable.id, o.assignedToUserId)).limit(1);
    assignedToUserName = u?.name ?? null;
  }
  return {
    id: o.id, source: o.source, status: o.status,
    sessionId: o.sessionId, assetId: o.assetId, assetName, assetNameAr,
    createdByUserId: o.createdByUserId, createdByUserName,
    assignedToUserId: o.assignedToUserId, assignedToUserName,
    customerName: o.customerName, totalAmount: parseFloat(o.totalAmount as string),
    cancelReason: o.cancelReason,
    createdAt: o.createdAt, preparingAt: o.preparingAt, readyAt: o.readyAt, deliveredAt: o.deliveredAt,
    items: items.map(i => ({
      ...i,
      unitPrice: parseFloat(i.unitPrice as string),
      totalPrice: parseFloat(i.totalPrice as string),
    })),
  };
}

router.get("/orders", requireAuth, requireTenant, async (req, res) => {
  try {
    const { status, source, sessionId } = req.query;
    let orders = await db.select().from(ordersTable)
      .where(eq(ordersTable.tenantId, req.user!.tenantId!))
      .orderBy(ordersTable.createdAt);
    if (status) orders = orders.filter(o => o.status === status);
    if (source) orders = orders.filter(o => o.source === source);
    if (sessionId) orders = orders.filter(o => o.sessionId === parseInt(sessionId as string));
    const result = await Promise.all(orders.reverse().slice(0, 200).map(buildOrderResponse));
    res.json(result);
  } catch {
    res.status(500).json({ error: "Failed to list orders" });
  }
});

router.get("/orders/kds", requireAuth, requireTenant, async (req, res) => {
  try {
    const orders = await db.select().from(ordersTable)
      .where(and(
        eq(ordersTable.tenantId, req.user!.tenantId!),
        inArray(ordersTable.status, ["pending", "preparing", "ready"])
      ))
      .orderBy(ordersTable.createdAt);
    const result = await Promise.all(orders.map(buildOrderResponse));
    res.json(result);
  } catch {
    res.status(500).json({ error: "Failed to list KDS orders" });
  }
});

router.get("/orders/:orderId", requireAuth, requireTenant, async (req, res) => {
  try {
    const id = parseInt(req.params.orderId);
    const [o] = await db.select().from(ordersTable)
      .where(and(eq(ordersTable.id, id), eq(ordersTable.tenantId, req.user!.tenantId!)))
      .limit(1);
    if (!o) { res.status(404).json({ error: "Not found" }); return; }
    res.json(await buildOrderResponse(o));
  } catch {
    res.status(500).json({ error: "Failed to get order" });
  }
});

async function createOrderWithItems(tenantId: number, data: {
  sessionId?: number; assetId?: number; source: string; customerName?: string;
  createdByUserId?: number; items: { productId: number; quantity: number; notes?: string }[];
}) {
  const { sessionId, assetId, source, customerName, createdByUserId, items } = data;
  // Get product prices
  const productIds = items.map(i => i.productId);
  const products = await db.select().from(productsTable)
    .where(and(eq(productsTable.tenantId, tenantId), inArray(productsTable.id, productIds)));
  const productMap = new Map(products.map(p => [p.id, p]));

  let totalAmount = 0;
  const itemsWithPrices = items.map(i => {
    const p = productMap.get(i.productId);
    if (!p) throw new Error(`Product ${i.productId} not found`);
    const unitPrice = parseFloat(p.price as string);
    const totalPrice = unitPrice * i.quantity;
    totalAmount += totalPrice;
    return { productId: i.productId, quantity: i.quantity, unitPrice: String(unitPrice), totalPrice: String(totalPrice), notes: i.notes };
  });

  const [order] = await db.insert(ordersTable).values({
    tenantId, sessionId, assetId, source, customerName,
    createdByUserId, totalAmount: String(totalAmount), status: "pending",
  }).returning();

  for (const item of itemsWithPrices) {
    await db.insert(orderItemsTable).values({ orderId: order.id, ...item });
  }
  return order;
}

router.post("/orders", requireAuth, requireTenant, async (req, res) => {
  try {
    const { sessionId, assetId, items, customerName } = req.body;
    if (!items?.length) { res.status(400).json({ error: "items required" }); return; }
    const order = await createOrderWithItems(req.user!.tenantId!, {
      sessionId, assetId, source: "pos",
      createdByUserId: req.user!.id, items, customerName,
    });
    await writeAuditLog({ user: req.user, action: "create_order", entityType: "order", entityId: order.id });
    res.status(201).json(await buildOrderResponse(order));
  } catch {
    res.status(500).json({ error: "Failed to create order" });
  }
});

router.patch("/orders/:orderId/status", requireAuth, requireTenant, async (req, res) => {
  try {
    const id = parseInt(req.params.orderId);
    const { status } = req.body;
    if (!status) { res.status(400).json({ error: "status required" }); return; }
    const updates: Partial<typeof ordersTable.$inferInsert> = { status };
    if (status === "preparing") updates.preparingAt = new Date();
    if (status === "ready") updates.readyAt = new Date();
    if (status === "delivered") {
      updates.deliveredAt = new Date();
      // Auto-deduct inventory
      const [o] = await db.select().from(ordersTable)
        .where(and(eq(ordersTable.id, id), eq(ordersTable.tenantId, req.user!.tenantId!))).limit(1);
      if (o) {
        const items = await db.select().from(orderItemsTable).where(eq(orderItemsTable.orderId, id));
        // Simple deduction: 1 unit per item per quantity (recipe system simplified)
        for (const item of items) {
          const invItems = await db.select().from(inventoryItemsTable)
            .where(eq(inventoryItemsTable.tenantId, req.user!.tenantId!));
          // Match by product if recipe exists (simplified: skip if no match)
        }
      }
    }
    const [updated] = await db.update(ordersTable).set(updates)
      .where(and(eq(ordersTable.id, id), eq(ordersTable.tenantId, req.user!.tenantId!)))
      .returning();
    if (!updated) { res.status(404).json({ error: "Not found" }); return; }
    await writeAuditLog({ user: req.user, action: `order_status_${status}`, entityType: "order", entityId: id });
    res.json(await buildOrderResponse(updated));
  } catch {
    res.status(500).json({ error: "Failed to update order status" });
  }
});

router.patch("/orders/:orderId/assign", requireAuth, requireTenant, async (req, res) => {
  try {
    const id = parseInt(req.params.orderId);
    const { userId } = req.body;
    const [updated] = await db.update(ordersTable).set({ assignedToUserId: userId })
      .where(and(eq(ordersTable.id, id), eq(ordersTable.tenantId, req.user!.tenantId!)))
      .returning();
    if (!updated) { res.status(404).json({ error: "Not found" }); return; }
    res.json(await buildOrderResponse(updated));
  } catch {
    res.status(500).json({ error: "Failed to assign order" });
  }
});

router.post("/orders/:orderId/cancel", requireAuth, requireTenant, async (req, res) => {
  try {
    const id = parseInt(req.params.orderId);
    const { reason } = req.body;
    if (!reason) { res.status(400).json({ error: "reason required" }); return; }
    const [updated] = await db.update(ordersTable).set({ status: "cancelled", cancelReason: reason })
      .where(and(eq(ordersTable.id, id), eq(ordersTable.tenantId, req.user!.tenantId!)))
      .returning();
    if (!updated) { res.status(404).json({ error: "Not found" }); return; }
    await writeAuditLog({ user: req.user, action: "cancel_order", entityType: "order", entityId: id, newValue: { reason } });
    res.json(await buildOrderResponse(updated));
  } catch {
    res.status(500).json({ error: "Failed to cancel order" });
  }
});

export { createOrderWithItems };
export default router;