# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Gaming Lounge OS API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /orders:
    get:
      operationId: listOrders
      tags: [orders]
      summary: List orders
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, preparing, ready, delivered, closed, cancelled]
        - name: source
          in: query
          schema:
            type: string
            enum: [qr, pos]
        - name: sessionId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
    post:
      operationId: createOrder
      tags: [orders]
      summary: Create a POS order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderInput"
      responses:
        "201":
          description: Order created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
...
  /orders/{orderId}:
    get:
      operationId: getOrder
      tags: [orders]
      summary: Get an order
      parameters:
        - name: orderId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
  /orders/{orderId}/status:
    patch:
      operationId: updateOrderStatus
      tags: [orders]
      summary: Update order status
      parameters:
        - name: orderId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderStatusUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
  /orders/{orderId}/assign:
    patch:
      operationId: assignOrder
      tags: [orders]
      summary: Assign order to a buffet worker
      parameters:
        - name: orderId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderAssignInput"
      responses:
        "200":
          description: Assigned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
  /orders/{orderId}/cancel:
    post:
      operationId: cancelOrder
      tags: [orders]
      summary: Cancel an order with reason
      parameters:
        - name: orderId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CancelInput"
      responses:
        "200":
          description: Cancelled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"

  # ── PAYMENTS ──────────────────────────────────────────
...
  /inventory:
    get:
      operationId: listInventoryItems
      tags: [inventory]
      summary: List inventory items
      responses:
        "200":
          description: Inventory items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/InventoryItem"
    post:
      operationId: createInventoryItem
      tags: [inventory]
      summary: Create inventory item
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InventoryItemInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InventoryItem"
  /inventory/{itemId}:
    patch:
      operationId: updateInventoryItem
      tags: [inventory]
      summary: Update inventory item
      parameters:
        - name: itemId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InventoryItemUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InventoryItem"
  /inventory/movements:
    get:
      operationId: listInventoryMovements
      tags: [inventory]
      summary: List inventory movements
      parameters:
        - name: itemId
          in: query
          schema:
            type: integer
        - name: type
          in: query
          schema:
            type: string
            enum: [sale, waste, adjustment, purchase]
      responses:
        "200":
          description: Movements
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/InventoryMovement"
    post:
      operationId: createInventoryMovement
      tags: [inventory]
      summary: Create inventory adjustment (manager approval required)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InventoryMovementInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InventoryMovement"
...
components:
  schemas:
...
    CancelInput:
      type: object
      required: [reason]
      properties:
        reason:
          type: string

    # ── PRODUCT SCHEMAS ──
...
    OrderItemInput:
      type: object
      required: [productId, quantity]
      properties:
        productId:
          type: integer
        quantity:
          type: integer
        notes:
          type: string

    # ── ORDER SCHEMAS ──
    OrderItem:
      type: object
      required: [id, productId, productName, quantity, unitPrice, totalPrice]
      properties:
        id:
          type: integer
        productId:
          type: integer
        productName:
          type: string
        productNameAr:
          type: ["string", "null"]
        quantity:
          type: integer
        unitPrice:
          type: number
        totalPrice:
          type: number
        notes:
          type: ["string", "null"]
    Order:
      type: object
      required: [id, source, status, totalAmount, items, createdAt]
      properties:
        id:
          type: integer
        source:
          type: string
          enum: [qr, pos]
        status:
          type: string
          enum: [pending, preparing, ready, delivered, closed, cancelled]
        sessionId:
          type: ["integer", "null"]
        assetId:
          type: ["integer", "null"]
        assetName:
          type: ["string", "null"]
        assetNameAr:
          type: ["string", "null"]
        createdByUserId:
          type: ["integer", "null"]
        createdByUserName:
          type: ["string", "null"]
        assignedToUserId:
          type: ["integer", "null"]
        assignedToUserName:
          type: ["string", "null"]
        customerName:
          type: ["string", "null"]
        totalAmount:
          type: number
        cancelReason:
          type: ["string", "null"]
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
        createdAt:
          type: string
          format: date-time
        preparingAt:
          type: ["string", "null"]
          format: date-time
        readyAt:
          type: ["string", "null"]
          format: date-time
        deliveredAt:
          type: ["string", "null"]
          format: date-time
    OrderInput:
      type: object
      required: [items]
      properties:
        sessionId:
          type: integer
        assetId:
          type: integer
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItemInput"
        customerName:
          type: string
    OrderStatusUpdate:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [pending, preparing, ready, delivered, closed, cancelled]
    OrderAssignInput:
      type: object
      required: [userId]
      properties:
        userId:
          type: integer

    # ── PAYMENT SCHEMAS ──
...
    InventoryItem:
      type: object
      required: [id, name, unit, currentStock, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        nameAr:
          type: ["string", "null"]
        unit:
          type: string
        currentStock:
          type: number
        minStockLevel:
          type: ["number", "null"]
        isLowStock:
          type: boolean
        createdAt:
          type: string
          format: date-time
    InventoryItemInput:
      type: object
      required: [name, unit, currentStock]
      properties:
        name:
          type: string
        nameAr:
          type: string
        unit:
          type: string
        currentStock:
          type: number
        minStockLevel:
          type: number
    InventoryItemUpdate:
      type: object
      properties:
        name:
          type: string
        nameAr:
          type: string
        unit:
          type: string
        currentStock:
          type: number
        minStockLevel:
          type: number
    InventoryMovement:
      type: object
      required: [id, inventoryItemId, type, quantity, createdAt]
      properties:
        id:
          type: integer
        inventoryItemId:
          type: integer
        itemName:
          type: ["string", "null"]
        type:
          type: string
          enum: [sale, waste, adjustment, purchase]
        quantity:
          type: number
        reason:
          type: ["string", "null"]
        approvedByUserId:
          type: ["integer", "null"]
        approvedByUserName:
          type: ["string", "null"]
        createdByUserId:
          type: ["integer", "null"]
        createdAt:
          type: string
          format: date-time
    InventoryMovementInput:
      type: object
      required: [inventoryItemId, type, quantity]
      properties:
        inventoryItemId:
          type: integer
        type:
          type: string
          enum: [waste, adjustment, purchase]
        quantity:
          type: number
        reason:
          type: string

    # ── DASHBOARD SCHEMAS ──
...